Reshuffled code order + more
I added a function to soundpairTable which allows for the creation of a joint table with both sounds.
This function checks to make sure each soundpair shares the same PIE origin.

I reshuffled the code. It now starts with the first word and goes through it letter by letter.
It checks each letter to see if it appears in the soundpair Table.
If it appears it then checks if the match is found in the second table.
If the match is found it takes note of where and searches after that for the next possible sound match.
It prints out the matches found.
It also prints out if it found a sound in one word but did not find its pair in the other.

I deleted cogpair as it was no longer needed.

I ensured all the strings used for the words and sounds were u32strings