Fixed issues regarding pʰ, tʰ, kʰ
Changed the way the first word is handled.
Instead of taking out an individual letter from the first word it is now checked with the function "longest_prefix"
This function is stored in soundpairTable. However This may be for organisation and neatness.
The function uses the sound table to check for the longest combination of letters in it that match with the word.
If found the matched letter/sound is then ignored and the function moves onto the next part of the word.
If not found the first char is ignored and the function moves onto the next part of the word.
This deals with the issues surrrounding pʰ etc, having more than one char.

Changed the way the second word is checked.
The change for the second word is much smaller.
The table is organised in order from largest to smallest sounds for the second language
And then based on that the biggest sound (in terms of ints) is checked first.
This means that pʰ will always be checked before p, stopping mistakes from happening.

Tested it against the current table of known cognates and it got all but one correct.
It does not see kerven and γράφω as cognates.
This is because none of the letters occuring in kerven are cognates according to my tables.
Will investigate.

Next step will be to upload a test with made up cognates and false cognates.
To test how the code reacts in difficult circumstances.